03.lambdas/02.connection_ssh_tunnel/requirements.txt
psycopg2-binary
paramiko
sshtunnel
python-dotenv
